move schema.rs file to server directory
