+++ src/model.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "server")]
-use crate::schema::penguin_encounter;
+use crate::server::schema::penguin_encounter;
 
 #[cfg(feature = "server")]
 use diesel::prelude::*;
